+++ CHANGELOG.md
@@ -11,10 +11,15 @@ All notable changes to the cavalier_contours crate will be documented in this fi
 
 ### Changed 🔧
 
+- Reimplemented `line_circle_intr` function for improved numerical stability.
+- Added epsilon parameter to `parametric_from_point` function to allow consistency with other
+  calculations using fuzzy epsilon values.
 - Added `#[inline]` attribute to all of the small Vector2 and base math functions.
 
 ### Fixed 🐛
 
+- Fixed boolean operation bug caused by `line_circle_intr` numerical stability problem
+  ([#23](https://github.com/jbuckmccready/cavalier_contours/issues/23)).
 - Fixed some of the doc comments around pline boolean operation types.
 - Fixed polyline find_intersects to use pos_equal_eps passed in options for querying bounding boxes
   ([#22](https://github.com/jbuckmccready/cavalier_contours/pull/22)).